% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslm.R
\name{bayeslm}
\alias{bayeslm}
\title{Bayesian linear models for cardinal paired comparison data}
\usage{
bayeslm(
  noitems,
  nocompars,
  scores,
  vars,
  xmu = zeros(noitems, 1),
  xvar = diag(noitems),
  a0 = 2,
  b0 = 1,
  Edges = c(1, 2, 1, 3, 2, 3),
  data = NULL,
  datatype = "simulated",
  prior = "conju",
  tol = 1e-08
)
}
\arguments{
\item{noitems}{The number of items for the paired comparison.}

\item{nocompars}{The number of comparisons among \code{noitems}.}

\item{scores}{The true scores or merits of \code{noitems}}

\item{vars}{The constant variance of \code{scores}.}

\item{xmu}{The normal prior mean of \code{noitems}.}

\item{xvar}{The prior covariance matrix of \code{noitems}.}

\item{a0}{The prior shape of inverse gamma distribution.}

\item{b0}{The prior scale of inverse gamma distribution.}

\item{Edges}{The edge set of \code{noitems} in the graph.}

\item{data}{The data containing the comparison outcomes; defalut is NULL.}

\item{datatype}{The type of data to analyze: "simulated" (the default) for simulated data,
"real" for a given real dataset.}

\item{prior}{The type of prior: "conju" (default) for conjugate prior, 
"semi-conju" for semi-conjugate prior, "flat" for flat prior, and "ref" for reference prior}

\item{tol}{The tolerance value to control near zero eigen values; default is 1e-08.}
}
\value{
The posterior mean, variance and posterior shape and scale of normal inverse
gamma distribution.
}
\description{
This function performs Bayesian inference for cardinal paired comparison
data. The methodology allows for doing Bayesian analysis on restricted
parameter space
}
\note{
%% ~~further notes~~
}
\examples{
## maybe something with simulated data...
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Prince P. Osei and Ori Davidov
}
