% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslmsamples.R
\name{bayeslmsamples}
\alias{bayeslmsamples}
\title{Draws from the posterior distribution for Bayesian linear models for 
paired comparison data}
\usage{
bayeslmsamples(
  iter,
  Kitems,
  Kcompars,
  Kscore,
  Kvars,
  Kmuprior,
  Kvarprior,
  Kedges,
  prior,
  data,
  type
)
}
\arguments{
\item{iter}{The number of posterior draws.}

\item{Kitems}{The number of items for the paired comparison.}

\item{Kcompars}{A vector of pairwise comparisons among \code{Kitems}.}

\item{Kscore}{The true scores or merits of \code{Kitems}}

\item{Kvars}{The true constant variance of \code{Kscore}.}

\item{Kmuprior}{The normal prior mean for \code{Kitems}.}

\item{Kvarprior}{The prior covariance matrix of \code{Kitems}.}

\item{Kedges}{The edge set of \code{Kitems} in the graph.}

\item{prior}{The type of prior: "conju" (default) for conjugate prior, 
"semi-conju" for semi-conjugate prior, "flat" for flat prior, and "ref" 
for reference prior}

\item{data}{he data containing the comparison outcomes; defalut is NULL.}

\item{type}{The type of data to analyze: "simulated" (the default) for simulated data,
"real" for a given real dataset.}
}
\value{
Posterior draws from \code{\link{bayeslm}}
}
\description{
Draws from the posterior distribution for Bayesian linear models for 
paired comparison data
}
\note{
%% ~~further notes~~
}
\examples{

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Prince P. Osei and Ori Davidov
}
