% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compagraph.R
\name{compagraph}
\alias{compagraph}
\title{The cardinal paired comparison graph}
\usage{
compagraph(
  noitems,
  nocompars,
  scores,
  vars = 1,
  data = NULL,
  datatype = "simulated"
)
}
\arguments{
\item{noitems}{The number of items for the paired comparison.}

\item{nocompars}{A vector of the number of pairwise comparisons among \code{noitems}.}

\item{scores}{The true scores or merits of \code{noitems}.}

\item{vars}{The constant variance of \code{scores}; default is 1.}

\item{data}{The data containing the comparison outcomes; defalut is NULL.}

\item{datatype}{The type of data to analyze: "simulated" (the default) for simulated data,
"real" for a given real dataset.}
}
\value{
A comparison graph with samples on edges.
%% @note %% ~~further notes~~
}
\description{
This function build the paired comparison graph or use a given paired comparison
outcomes to build the graph (G,S).
}
\examples{
## K = 3 # number of items
## compars <- rep(3,3) # number of comparison (1,2,1,3,2,3)
## scores <-  1:3 # the true scores of the items
## tvar <- 1
## compagraph(K,compars,scores,tvar,data=NULL)
}
\author{
Prince P. Osei and Ori Davidov
%% @seealso %% ~~objects to See Also as \code{\link{help}}, ~~~
%% @references %% ~put references to the literature/web site here ~
}
