% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcbayeslm.R
\name{cpcbayeslm}
\alias{cpcbayeslm}
\title{Bayesian linear models for cardinal paired comparison data}
\usage{
cpcbayeslm(
  noitems,
  nocompars,
  scores,
  vars = 1,
  xmu = zeros(noitems, 1),
  xvar = vars * diag(noitems),
  a0 = 2,
  b0 = 1,
  Edges = c(1, 2, 1, 3, 2, 3),
  data = NULL,
  datatype = "simulated",
  prior = "conju",
  tol = 1e-08
)
}
\arguments{
\item{noitems}{The number of items for the paired comparison.}

\item{nocompars}{The number of comparisons among \code{noitems}.}

\item{scores}{The true scores or merits of \code{noitems}}

\item{vars}{The constant variance of \code{scores}; default is 1.}

\item{xmu}{The normal prior mean of \code{noitems}; default is zero vector of 
length \code{noitems}.}

\item{xvar}{The prior covariance matrix of \code{noitems}; default is identity matrix 
of dimension \code{noitems}.}

\item{a0}{The prior shape of inverse gamma distribution; default is 2.}

\item{b0}{The prior scale of inverse gamma distribution; default is 1.}

\item{Edges}{The edge set of \code{noitems} in the graph; default is edgeset of
three items.}

\item{data}{The data containing the comparison outcomes; defalut is NULL.}

\item{datatype}{The type of data to analyze: "simulated" (the default) for simulated data,
"real" for a given real dataset.}

\item{prior}{The type of prior: "conju" (default) for conjugate prior, 
"semi-conju" for semi-conjugate prior, "flat" for flat prior, and "ref" for reference prior}

\item{tol}{The tolerance value to control near zero eigen values; default is 1e-08.}
}
\value{
The Bayesian posterior mean and variance together with the least squares estimates.
%% @note %% ~~further notes~~
}
\description{
This function performs Bayesian inference for cardinal paired comparison
data. The methodology allows for doing Bayesian analysis on restricted
parameter space
}
\examples{
K = 3 # number of items
paircompars <- rep(3,3) # number of pairwise comparisons
Tscores <- 3:1-mean(3:1) # true scores 
cpcbayeslm(K,paircompars,Tscores)
}
\references{
Osei, P. P. and Davidov, O. (2022). Bayesian linear models for cardinal paired comparison data. Under revision.
}
\author{
Prince P. Osei and Ori Davidov
%% @seealso %% ~~objects to See Also as \code{\link{help}}, ~~~
}
