\name{internal}


\alias{funcrank}

\alias{zeros}

\alias{ones}

\alias{meritlabels}

\alias{vedges}

\alias{pathgraph}

\alias{edgescompars}

\alias{incompletegraph}

\alias{postranksfreq}

\alias{postprobsucra}


%- Also NEED an '\alias' for EACH other topic documented here.

\title{

Internal functions

}

\description{

These functions are ordinarily not to be called by the user, but if you contact the package authors with any questions about them, we'll do our best to clarify matters.}









