% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postdrawspcd.R
\name{postdrawspcd}
\alias{postdrawspcd}
\title{Draws samples from the posterior distribution for paired comparison data}
\usage{
postdrawspcd(
  N,
  noitems,
  nocompars,
  scores,
  vars = 1,
  xmu = zeros(noitems, 1),
  xvar = vars * diag(noitems),
  a0 = 2,
  b0 = 1,
  Edges = c(1, 2, 1, 3, 2, 3),
  data = NULL,
  datatype = "simulated",
  prior = "conju"
)
}
\arguments{
\item{N}{Number of posterior samples.}

\item{noitems}{Number of items in the comparison graph}

\item{nocompars}{The number of comparisons among \code{noitems}.}

\item{scores}{The true scores or merits of \code{noitems}}

\item{vars}{The constant variance of \code{scores} default is 1.}

\item{xmu}{The normal prior mean of \code{noitems} default is a vector of zeros of length
\code{noitems}.}

\item{xvar}{The prior covariance matrix of \code{noitems} default is \code{vars} times 
the identity matrix dimension \code{noitems}.}

\item{a0}{The prior shape of inverse gamma distribution default is 2}

\item{b0}{The prior scale of inverse gamma distribution default is 1}

\item{Edges}{The edge set of \code{noitems} in the graph default is complete graph of 
three \code{noitems}.}

\item{data}{The data containing the comparison outcomes; defalut is NULL.}

\item{datatype}{The type of data to analyze: "simulated" (the default) for simulated data,
"real" for a given real dataset.}

\item{prior}{The type of prior: "conju" (default) for conjugate prior, 
"semi-conju" for semi-conjugate prior, "flat" for flat prior, and "ref" for reference prior}
}
\value{
A posterior rankings of posterior draws from \code{\link{cpcbayeslm}}
%% @note %% ~~further notes~~
}
\description{
This function performs a new posterior ranking methodology developed in the 
paper by Prince P. Osei and Ori David (2021) to account for the uncertainty
in the ranking of items.
}
\examples{
##  A graph of three items
##  K <- 3 # number of items
## edges <- c(1,2,1,3,2,3) # edge set
## compars <- rep(3,3) # pairwise comparisons
##  Tscores <- 1:3-mean(1:3) # The true score sum to zero
##  N <- 100 # number of samples to draw
## postdrawspcd(N,K,compars,Tscores,Edges=edges)
}
\author{
Prince P. Osei and Ori Davidov
%% @seealso %% ~~objects to See Also as \code{\link{help}}, ~~~
%% @references %% ~put references to the literature/web site here ~
}
