% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projbayeslm.R
\name{projbayeslm}
\alias{projbayeslm}
\title{Projection of non-singular prior onto the subspace of the Laplacian}
\usage{
projbayeslm(noitems, xmu, xvar)
}
\arguments{
\item{noitems}{The number of items for the paired comparison.}

\item{xmu}{The normal prior mean of \code{noitems}.}

\item{xvar}{The prior covariance matrix of \code{noitems}.}
}
\value{
Projected mean and variance matrix of the prior distribution.
%% @note %% ~~further notes~~
}
\description{
Projection of non-singular prior onto the subspace of the Laplacian
}
\examples{
## K = 6 # number of items
## mu = rep(0,6) # the prior mean
## xcov = diag(6) # the prior covariance matrix
## projbayeslm(K,mu,xcov)
}
\author{
Prince P. Osei and Ori Davidov
%% @seealso %% ~~objects to See Also as \code{\link{help}}, ~~~
%% @references %% ~put references to the literature/web site here ~
}
