% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslmsamples.R
\name{bayeslmsamples}
\alias{bayeslmsamples}
\title{Draws from the posterior distribution for Bayesian linear models for 
paired comparison data}
\usage{
bayeslmsamples(
  iter,
  noitems,
  nocompars,
  scores,
  vars = 1,
  muprior = zeros(noitems, 1),
  varprior = vars * diag(noitems),
  Edges,
  prior = "conju",
  data = NULL,
  type = "simulated"
)
}
\arguments{
\item{iter}{The number of posterior draws.}

\item{noitems}{The number of items for the paired comparison.}

\item{nocompars}{A vector of pairwise comparisons among \code{noitems}.}

\item{scores}{The true scores or merits of \code{noitems}}

\item{vars}{The true constant variance of \code{scores}; default is 1.}

\item{muprior}{The normal prior mean for \code{noitems}; default is a vector of 
zeros of length \code{noitems}.}

\item{varprior}{The prior covariance matrix of \code{noitems}; default is the 
identity matrix of dimension \code{noitems}.}

\item{Edges}{The edge set of \code{noitems} in the graph.}

\item{prior}{The type of prior: "conju" (default) for conjugate prior, 
"semi-conju" for semi-conjugate prior, "flat" for flat prior, and "ref" 
for reference prior}

\item{data}{he data containing the comparison outcomes; defalut is NULL.}

\item{type}{The type of data to analyze: "simulated" (the default) for simulated data,
"real" for a given real dataset.}
}
\value{
Posterior draws from \code{\link{cpcbayeslm}}
%% @note %% ~~further notes~~
}
\description{
Draws from the posterior distribution for Bayesian linear models for 
paired comparison data
}
\examples{
## May be something from simulated data??
}
\author{
Prince P. Osei and Ori Davidov
%% @seealso %% ~~objects to See Also as \code{\link{help}}, ~~~
%% @references %% ~put references to the literature/web site here ~
}
