% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probranks.R
\name{probranks}
\alias{probranks}
\title{Ranks samples from posterior distribution of
Bayes linear models for paired comparison data}
\usage{
probranks(samples, ascending = T)
}
\arguments{
\item{samples}{The posterior draws obtained from \code{\link{cpcbayeslm}} and 
\code{\link{postdrawspcd}}}

\item{ascending}{A logical variable default to True for ascending order of the 
item labels.}
}
\value{
Ranks permutations and ranking probabilities
%% @note %% ~~further notes~~
}
\description{
Ranks samples from posterior distribution of
Bayes linear models for paired comparison data
}
\examples{
##  A graph of three items
##  K <- 3 # number of items
## edges <- c(1,2,1,3,2,3) # edge set
## compars <- rep(3,3) # pairwise comparisons
##  Tscores <- 1:3-mean(1:3) # The true score sum to zero
##  N <- 100 # number of samples to draw
## hh <- postdrawspcd(N,K,compars,Tscores,Edges=edges) # posterior draws
## probranks(hh$drawspost)
}
\author{
Prince P. Osei and Ori Davidov
%% @seealso %% ~~objects to See Also as \code{\link{help}}, ~~~
%% @references %% ~put references to the literature/web site here ~
}
