% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probRanks.R
\name{probRanks}
\alias{probRanks}
\title{Ranks samples from Bayes linear models for paired comparison data}
\usage{
probRanks(samples, ascending = T)
}
\arguments{
\item{samples}{The posterior draws obtained from \code{\link{bayeslm}}}

\item{ascending}{A logical variable default to True for ascending order of the 
item labels.}
}
\value{
Ranks permutations and ranking probabilities
}
\description{
Ranks samples from Bayes linear models for paired comparison data
}
\note{
%% ~~further notes~~
}
\examples{
## May be simulated data??
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Prince P. Osei and Ori Davidov
}
